Move common code to parent class
