+++ app/app.rb
@@ -1,8 +1,11 @@
+# frozen_string_literal: true
+
 require 'nokogiri'
 require 'pry'
 require 'json'
 require 'rspec'
 
 require_relative './parsers/nokogiri_html_parser'
+require_relative './scrapers/google_carousel'
 require_relative './scrapers/google_carousel_20190319'
 require_relative './scrapers/google_carousel_20230813'
