+++ spec/scrapers/google_carousel_20190319_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 RSpec.describe Scrapers::GoogleCarousel20190319 do
