+++ spec/scrapers/google_carousel_20230813_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 RSpec.describe Scrapers::GoogleCarousel20230813 do
